Fix warning in Add benchmark

+++ bench/add.cc
@@ -138,7 +138,7 @@ static void CharacteristicArguments(benchmark::internal::Benchmark* b)
 
   uint32_t c = 16;
   for (uint32_t n = 224; n >= 7; n /= 2) {
-    b->Args({n * n, c});
+    b->Args({int32_t(n * n), c});
     c *= 2;
   }
 }
